+++ src/main/java/lermitage/intellij/extra/icons/UITypeIconsPreference.java
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT
+
+package lermitage.intellij.extra.icons;
+
+public enum UITypeIconsPreference {
+    BASED_ON_ACTIVE_UI_TYPE,
+    PREFER_NEW_UI_ICONS,
+    PREFER_OLD_UI_ICONS
+}