+++ src/main/java/lermitage/intellij/extra/icons/utils/IconUtils.java
@@ -14,6 +14,7 @@ import com.intellij.util.ui.ImageUtil;
 import com.intellij.util.ui.JBImageIcon;
 import lermitage.intellij.extra.icons.IconType;
 import lermitage.intellij.extra.icons.Model;
+import lermitage.intellij.extra.icons.UITypeIconsPreference;
 import lermitage.intellij.extra.icons.cfg.services.SettingsService;
 import org.jetbrains.annotations.NonNls;
 
@@ -33,16 +34,29 @@ public class IconUtils {
 
     private static final @NonNls Logger LOGGER = Logger.getInstance(IconUtils.class);
 
+    private static final UITypeIconsPreference UI_TYPE_ICONS_PREFERENCE = SettingsService.getIDEInstance().getUiTypeIconsPreference();
+
     private static final int SCALING_SIZE = 16;
+
     //private static final Pattern cssVarRe = Pattern.compile("var\\([-\\w]+\\)");
 
     public static Icon getIcon(Model model, double additionalUIScale) {
         if (model.getIconType() == IconType.PATH) {
-            if (UIUtils.isNewUIEnabled() && model.isAutoLoadNewUIIconVariant()) {
-                String newUIIconPath = "/" + model.getIcon().replace("extra-icons/", "extra-icons/newui/"); //NON-NLS
-                return IconLoader.getIcon(newUIIconPath, IconUtils.class); //NON-NLS
+            // use the new UI icon if exists + asked by user + new UI enabled, otherwise use the old UI icon
+            String iconPathToLoad = model.getIcon(); // defaults to old UI icon
+            switch (UI_TYPE_ICONS_PREFERENCE) {
+                case BASED_ON_ACTIVE_UI_TYPE -> {
+                    if (UIUtils.isNewUIEnabled() && model.isAutoLoadNewUIIconVariant()) {
+                        iconPathToLoad = "/" + model.getIcon().replace("extra-icons/", "extra-icons/newui/"); //NON-NLS
+                    }
+                }
+                case PREFER_NEW_UI_ICONS -> {
+                    if (model.isAutoLoadNewUIIconVariant()) {
+                        iconPathToLoad = "/" + model.getIcon().replace("extra-icons/", "extra-icons/newui/"); //NON-NLS
+                    }
+                }
             }
-            return IconLoader.getIcon(model.getIcon(), IconUtils.class);
+            return IconLoader.getIcon(iconPathToLoad, IconUtils.class);
         }
         ImageWrapper fromBase64 = fromBase64(model.getIcon(), model.getIconType(), additionalUIScale);
         if (fromBase64 == null) {