+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2023.2.7-231 (WIP)
 * support [Just](https://github.com/casey/just) `justfile` files.
+* you can now choose to use new UI icon variants based on current active UI, or force the use of old or new UI icons, when variants exist. See Extra Icons settings.
 * add *new UI* icons for JPA Buddy, Flyway, Liquibase and TeamCity folders. New UI icons are automatically loaded when needed.
 * add *new UI* icons for various Helm files and folders.
 * add *new UI* icons for various archive files.